+++ config/index.js
@@ -195,10 +195,10 @@ module.exports = {
     server_address: process.env.BLOT_ICLOUD_SERVER_ADDRESS,
     email: process.env.BLOT_ICLOUD_EMAIL,
     // The maximum file size to sync with iCloud
-    maxFileSize: 1e6, // 1MB
+    maxFileSize: 1e8, // 100MB
     // The thresholds for sending warning emails
-    diskSpaceWarning: 5e9, // 5GB
-    diskSpaceLimit: 1e9, // 1GB
+    diskSpaceWarning: 1e10, // 10GB
+    diskSpaceLimit: 5e9, // 5GB
     iCloudSpaceWarning: 2e9, // 2GB
     iCloudSpaceLimit: 1e9, // 1GB
   }
